Move GetAuthConfigFromPullSecretPath to shared/utils and cleanup dead code
Moves the credential functions out of the config package and into
shared/pkg/utils/creds, and also removes all the now unneeded functions
left behind by the change.
