+++ shared/pkg/utils/k8s/k8s_utils_test.go
@@ -17,9 +17,6 @@ package k8s
 
 import (
 	"testing"
-
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 func TestParseImageHash(t *testing.T) {
@@ -57,187 +54,6 @@ func TestParseImageHash(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			if got := ParseImageHash(tt.args.imageID); got != tt.want {
 				t.Errorf("ParseImageHash() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestGetMatchingSecretName(t *testing.T) {
-	notSpecificImageName := "gcr.io/not/specific"
-	specificImageName := "gcr.io/more/specific"
-	partialPathsImageName := "gcr.io/partial/path"
-	dockerHubImageName := "repo/image"
-
-	dockerconfigjsonB := []byte("{\"auths\":{\"gcr.io\":{\"username\":\"gcr\",\"password\":\"io\",\"auth\":\"Z2NyOmlv\"},\"gcr.io/more/specific\":{\"username\":\"gcr\",\"password\":\"io/more/specific\",\"auth\":\"Z2NyOmlvL21vcmUvc3BlY2lmaWM=\"},\"http://foo.example.com\":{\"username\":\"foo\",\"password\":\"bar\",\"auth\":\"Zm9vOmJhcg==\"}}}")
-	partialPathsDockerconfigjsonB := []byte("{\"auths\":{\"gcr.io/partial\":{\"username\":\"gcr\",\"password\":\"partial\",\"auth\":\"Z2NyOmlvL21vcmUvc3BlY2lmaWM=\"},\"http://foo.example.com\":{\"username\":\"foo\",\"password\":\"bar\",\"auth\":\"Zm9vOmJhcg==\"}}}")
-	noMatchDockerconfigjsonB := []byte("{\"auths\":{\"http://foo.example.com\":{\"username\":\"foo\",\"password\":\"bar\",\"auth\":\"Zm9vOmJhcg==\"}}}")
-	matchDockerIODockerconfigjsonB := []byte("{\"auths\":{\"https://index.docker.io/v1/\":{\"username\":\"test-user\",\"password\":\"test-pass)\",\"email\":\"test@test\",\"auth\":\"dGVzdC11c2VyOnRlc3QtcGFzcw==\"}}}")
-
-	type args struct {
-		secrets   []*corev1.Secret
-		imageName string
-	}
-	tests := []struct {
-		name string
-		args args
-		want string
-	}{
-		{
-			name: "match specific image name",
-			args: args{
-				secrets: []*corev1.Secret{
-					{
-						TypeMeta: metav1.TypeMeta{
-							Kind:       "Secret",
-							APIVersion: "v1",
-						},
-						ObjectMeta: metav1.ObjectMeta{
-							Name:      "regcred",
-							Namespace: "default",
-						},
-						Data: map[string][]byte{
-							corev1.DockerConfigJsonKey: dockerconfigjsonB,
-						},
-						Type: corev1.SecretTypeDockerConfigJson,
-					},
-				},
-				imageName: specificImageName,
-			},
-			want: "regcred",
-		},
-		{
-			name: "match partial paths image name",
-			args: args{
-				secrets: []*corev1.Secret{
-					{
-						TypeMeta: metav1.TypeMeta{
-							Kind:       "Secret",
-							APIVersion: "v1",
-						},
-						ObjectMeta: metav1.ObjectMeta{
-							Name:      "regcred",
-							Namespace: "default",
-						},
-						Data: map[string][]byte{
-							corev1.DockerConfigJsonKey: partialPathsDockerconfigjsonB,
-						},
-						Type: corev1.SecretTypeDockerConfigJson,
-					},
-				},
-				imageName: partialPathsImageName,
-			},
-			want: "regcred",
-		},
-		{
-			name: "match not specific image name",
-			args: args{
-				secrets: []*corev1.Secret{
-					{
-						TypeMeta: metav1.TypeMeta{
-							Kind:       "Secret",
-							APIVersion: "v1",
-						},
-						ObjectMeta: metav1.ObjectMeta{
-							Name:      "no-match",
-							Namespace: "default",
-						},
-						Data: map[string][]byte{
-							corev1.DockerConfigJsonKey: noMatchDockerconfigjsonB,
-						},
-						Type: corev1.SecretTypeDockerConfigJson,
-					},
-					{
-						TypeMeta: metav1.TypeMeta{
-							Kind:       "Secret",
-							APIVersion: "v1",
-						},
-						ObjectMeta: metav1.ObjectMeta{
-							Name:      "regcred",
-							Namespace: "default",
-						},
-						Data: map[string][]byte{
-							corev1.DockerConfigJsonKey: dockerconfigjsonB,
-						},
-						Type: corev1.SecretTypeDockerConfigJson,
-					},
-				},
-				imageName: notSpecificImageName,
-			},
-			want: "regcred",
-		},
-		{
-			name: "no match",
-			args: args{
-				secrets: []*corev1.Secret{
-					{
-						TypeMeta: metav1.TypeMeta{
-							Kind:       "Secret",
-							APIVersion: "v1",
-						},
-						ObjectMeta: metav1.ObjectMeta{
-							Name:      "no-match",
-							Namespace: "default",
-						},
-						Data: map[string][]byte{
-							corev1.DockerConfigJsonKey: noMatchDockerconfigjsonB,
-						},
-						Type: corev1.SecretTypeDockerConfigJson,
-					},
-				},
-				imageName: notSpecificImageName,
-			},
-			want: "",
-		},
-		{
-			name: "not docker config json secret",
-			args: args{
-				secrets: []*corev1.Secret{
-					{
-						TypeMeta: metav1.TypeMeta{
-							Kind:       "Secret",
-							APIVersion: "v1",
-						},
-						ObjectMeta: metav1.ObjectMeta{
-							Name:      "no-docker-config",
-							Namespace: "default",
-						},
-						Data: map[string][]byte{
-							"bla": []byte("bla"),
-						},
-					},
-				},
-				imageName: notSpecificImageName,
-			},
-			want: "",
-		},
-		{
-			name: "match docker hub image",
-			args: args{
-				secrets: []*corev1.Secret{
-					{
-						TypeMeta: metav1.TypeMeta{
-							Kind:       "Secret",
-							APIVersion: "v1",
-						},
-						ObjectMeta: metav1.ObjectMeta{
-							Name:      "regcred",
-							Namespace: "default",
-						},
-						Data: map[string][]byte{
-							corev1.DockerConfigJsonKey: matchDockerIODockerconfigjsonB,
-						},
-						Type: corev1.SecretTypeDockerConfigJson,
-					},
-				},
-				imageName: dockerHubImageName,
-			},
-			want: "regcred",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := GetMatchingSecretName(tt.args.secrets, tt.args.imageName); got != tt.want {
-				t.Errorf("GetMatchingSecretName() = %v, want %v", got, tt.want)
 			}
 		})
 	}
